refactor(windows): extract settings and config transfer into funput-config
`platforms/windows/src` kept ~900 lines of pure serde behind a crate that only
builds on Windows and needs a full skia build to compile at all, so none of it
could be exercised from a developer machine on another OS. Move the settings
model, the path policy, and the Export/Import document into a new workspace
crate; the Windows shell keeps only what needs `current_exe`/`config_dir`
(`settings_path.rs`).

25 tests now run under `cargo test --workspace` for the first time, and clippy
reaches this code for the first time too — it immediately flagged a
`collapsible_if` that had been sitting in `storage.rs` unseen.

Two things are not a straight move:

- `to_document` now takes `Source` from the caller. It used to read
  `env!("CARGO_PKG_VERSION")`, which after the move would have written the
  config crate's version (0.1.0) into every exported file instead of the app's.
- The temp-dir helper the moved tests share now uses an atomic counter. Naming
  by timestamp alone collides when `cargo test` runs them on parallel threads,
  and one test's cleanup then deletes another's directory.

The on-disk format is unchanged: every serde attribute moved verbatim, and new
tests pin the camelCase key names, a full-file read, and a legacy file that
predates the optional fields.

Settings is deliberately a superset of the Linux model and `Platform` has room
for a `linux` block, so `platforms/linux/settings-gtk` — which reimplements all
of this today — can adopt the crate later. Not migrated here: this machine
cannot build gtk4/libadwaita to verify it.
