+++ crates/funput-config/src/settings/combo.rs
