+++ platforms/windows/src/commands.rs
@@ -5,8 +5,8 @@
 use std::path::Path;
 use std::sync::Mutex;
 
-use crate::config_transfer::{self, ConfigError, ImportSummary};
-use crate::settings::{ExcludedApp, FlipHotkey, Hotkey, KeyCombo, Method, ToneStyle};
+use funput_config::transfer::{self, ConfigError, ImportSummary, Source};
+use funput_config::{ExcludedApp, FlipHotkey, Hotkey, KeyCombo, Method, ToneStyle};
 use crate::update::{self, Manifest};
 use crate::{shell, windows_ui};
 
@@ -123,14 +123,21 @@ pub fn open_url(url: &str) {
 // --- Config Export / Import -------------------------------------------------
 
 /// Write the current settings to `path` as an interchange file.
+///
+/// `Source` is filled in here, not by `funput-config`: `CARGO_PKG_VERSION` has to
+/// be *this* crate's version (the shipped app's), not the config crate's.
 pub fn export_config(path: &Path) -> std::io::Result<()> {
-    config_transfer::export_to(path, &shell::snapshot())
+    let source = Source {
+        platform: "windows".to_string(),
+        app_version: env!("CARGO_PKG_VERSION").to_string(),
+    };
+    transfer::export_to(path, &shell::snapshot(), source)
 }
 
 /// Merge a config file into the live settings (applies to the engine + persists).
 pub fn import_config(path: &Path) -> Result<ImportSummary, ConfigError> {
     let mut settings = shell::snapshot();
-    let summary = config_transfer::import_file(path, &mut settings)?;
+    let summary = transfer::import_file(path, &mut settings)?;
     shell::replace_settings(settings);
     Ok(summary)
 }
