+++ platforms/windows/src/main.rs
@@ -12,7 +12,6 @@ slint::include_modules!();
 mod canonical_exe;
 mod commands;
 mod compose;
-mod config_transfer;
 mod dark_mode;
 mod hook;
 mod hotkey;
@@ -21,7 +20,7 @@ mod instance;
 mod keymap;
 mod mica;
 mod recorder;
-mod settings;
+mod settings_path;
 mod shell;
 mod system_accent;
 mod tray;
